update order controller to check for user when editing orders

+++ app/Http/Controllers/OrdersController.php
@@ -2,10 +2,11 @@
 
 namespace App\Http\Controllers;
 
-use Illuminate\Http\Request;
-use App\Http\Controllers\Controller;
 use App\Models\Orders;
+use Illuminate\Http\Request;
 use App\Services\OrdersService;
+use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Auth;
 
 class OrdersController extends Controller
 {
@@ -17,30 +18,34 @@ class OrdersController extends Controller
 
     public function index()
     {
-        $orders = [];
-        foreach (Orders::all() as $order) {
-            $orders[] = $order;
-        }
-
+        $user = Auth::user();
+        $orders = Orders::where('id', $user->id)->get();
         return view('orders.index', ['orders' => $orders]);
     }
 
-
     public function store(Request $request)
     {
-        $openOrder = Orders::where('state', 'OPEN')->first();
+        $user = Auth::user();
+        $openOrder = Orders::where('state', 'OPEN')
+            ->where('id', $user->id)
+            ->first();
         if ($openOrder) {
             app(OrdersService::class)->update($openOrder, $request);
-            return redirect()->route('dashboard');
         } else {
             app(OrdersService::class)->create($request);
-            return redirect()->route('dashboard');
         }
+        return redirect()->route('dashboard');
     }
 
     public function delete_class(Request $request)
     {
-        $order = Orders::find($request->order_id);
-        app(OrdersService::class)->delete($order, $request);
+        $user = Auth::user();
+        $order = Orders::where('id', $request->order_id)
+            ->where('id', $user->id)
+            ->first();
+        if ($order) {
+            app(OrdersService::class)->delete($order, $request);
+        }
+        return redirect()->route('dashboard');
     }
 }
